+++ slices/collect_test.go
@@ -3,6 +3,8 @@ package slices_test
 import (
 	"github.com/synesissoftware/ANGoLS/slices"
 
+	"github.com/stretchr/testify/assert"
+
 	"fmt"
 	"strconv"
 	"strings"
@@ -10,48 +12,50 @@ import (
 )
 
 const (
-	int_equal_iterations = 100000
-	int_equal_size       = 100
+	int_equal_size = 1_000
 )
 
 // Tests
 
+func Test_CollectSlice_int_INTO_string(t *testing.T) {
+
+	collector := func(index int, input_item *int) (string, error) {
+
+		return strconv.Itoa(*input_item), nil
+	}
+
+	input := []int{1, -2, 3}
+	expected := []string{"1", "-2", "3"}
+
+	actual, err := slices.CollectSlice(input, collector)
+	if err != nil {
+
+	} else {
+		assert.Equal(t, expected, actual)
+	}
+}
+
 func Test_Collect_Array_1_ints_to_ints(t *testing.T) {
 
-	fn := func(input any) (any, error) {
+	collector := func(index int, input_item *int) (int, error) {
 
-		if i, ok := input.(int); ok {
+		if *input_item < 0 {
 
-			if i < 0 {
-
-				return -i, nil
-			}
+			return -*input_item, nil
 		}
 
-		return input, nil
+		return *input_item, nil
 	}
 
 	ints := []int{1, 2, 3}
 
-	r, err := slices.CollectSlice(ints, fn)
+	r, err := slices.CollectSlice(ints, collector)
 	if err != nil {
 
 		t.Errorf("Collect() failed: %v", err)
 	} else {
 
-		for ix, v := range r {
-
-			if i, ok := v.(int); !ok {
-
-				t.Errorf("Element (%T)%v of non-int type at index %d\n", v, v, ix)
-			} else {
-
-				if ints[ix] != i {
-
-					t.Errorf("Element (%T)%v at index %d is different from expected value %d\n", v, v, ix, ints[ix])
-				}
-			}
-		}
+		assert.Equal(t, ints, r)
 	}
 }
 
@@ -120,7 +124,7 @@ func Test_CollectSliceOfInteger_WITH_uint(t *testing.T) {
 	expected := []uint{2, 0, 4, 6}
 	actual := slices.CollectSliceOfInteger(input_slice, fn)
 
-	if !slices.EqualSliceOfUInt(expected, actual) {
+	if !slices.EqualSliceOfUint(expected, actual) {
 
 		t.Errorf("actual value '%v' does not equal expected value '%v'", actual, expected)
 	}
@@ -183,7 +187,9 @@ func Benchmark_equal_ints_by_EqualSliceOfInt(b *testing.B) {
 	ints_1, _ := slices.GenerateSliceOfInt(int_equal_size, fn)
 	ints_2, _ := slices.GenerateSliceOfInt(int_equal_size, fn)
 
-	for i := 0; i != int_equal_iterations; i++ {
+	b.ResetTimer()
+
+	for b.Loop() {
 
 		_ = slices.EqualSliceOfInt(ints_1, ints_2)
 	}
@@ -195,7 +201,9 @@ func Benchmark_equal_ints_by_EqualSlice(b *testing.B) {
 	ints_1, _ := slices.GenerateSliceOfInt(int_equal_size, fn)
 	ints_2, _ := slices.GenerateSliceOfInt(int_equal_size, fn)
 
-	for i := 0; i != int_equal_iterations; i++ {
+	b.ResetTimer()
+
+	for b.Loop() {
 
 		_ = slices.EqualSlice(ints_1, ints_2)
 	}
@@ -204,10 +212,12 @@ func Benchmark_equal_ints_by_EqualSlice(b *testing.B) {
 func Benchmark_equal_uints_by_EqualSlice(b *testing.B) {
 
 	fn := func(index int) (uint, error) { return uint(index), nil }
-	ints_1, _ := slices.GenerateSliceOfUInt(int_equal_size, fn)
-	ints_2, _ := slices.GenerateSliceOfUInt(int_equal_size, fn)
+	ints_1, _ := slices.GenerateSliceOfUint(int_equal_size, fn)
+	ints_2, _ := slices.GenerateSliceOfUint(int_equal_size, fn)
 
-	for i := 0; i != int_equal_iterations; i++ {
+	b.ResetTimer()
+
+	for b.Loop() {
 
 		_ = slices.EqualSlice(ints_1, ints_2)
 	}
@@ -223,7 +233,7 @@ func Test_CollectSliceIntoStringSlice(t *testing.T) {
 		"Maverick",
 	}
 
-	result, err := slices.CollectSliceIntoStringSlice[Nickname](nicknames, func(input_item *Nickname) (string, error) {
+	result, err := slices.CollectSliceIntoStringSlice(nicknames, func(input_item *Nickname) (string, error) {
 		return string(*input_item), nil
 	})
 
@@ -248,7 +258,7 @@ func Test_CollectSliceIntoStringSlice_with_CaseTesting(t *testing.T) {
 		"Maverick",
 	}
 
-	result, err := slices.CollectSliceIntoStringSlice[Nickname](nicknames, func(input_item *Nickname) (string, error) {
+	result, err := slices.CollectSliceIntoStringSlice(nicknames, func(input_item *Nickname) (string, error) {
 		return strings.ToUpper(string(*input_item)), nil
 	})
 
@@ -271,7 +281,7 @@ func Test_CollectSliceIntoStringSlice_with_Ints(t *testing.T) {
 		3,
 	}
 
-	result, err := slices.CollectSliceIntoStringSlice[int](values, func(input_item *int) (string, error) {
+	result, err := slices.CollectSliceIntoStringSlice(values, func(input_item *int) (string, error) {
 		return strconv.Itoa(*input_item), nil
 	})
 
@@ -296,7 +306,7 @@ func Test_CollectSliceIntoStringSlice_with_Fail(t *testing.T) {
 		"Maverick",
 	}
 
-	_, err := slices.CollectSliceIntoStringSlice[Nickname](nicknames, func(input_item *Nickname) (string, error) {
+	_, err := slices.CollectSliceIntoStringSlice(nicknames, func(input_item *Nickname) (string, error) {
 		if strings.Contains(string(*input_item), "k") {
 			return "", fmt.Errorf("invalid nickname: %s", *input_item)
 		}