+++ strings/ascii.go
@@ -4,7 +4,7 @@
 
 /*
  * Created: 1st January 2026
- * Updated: 2nd January 2026
+ * Updated: 7th February 2026
  */
 
 package strings
@@ -17,12 +17,17 @@ package strings
 // are known to contain only ASCII and a faster conversion than is provided
 // by the standard `ToLower()` is desired.
 func ASCIIToLower(s string) string {
+
 	b := []byte(s)
-	for i := 0; i < len(b); i++ {
-		if b[i] >= 'A' && b[i] <= 'Z' {
+
+	for i, c := range b {
+
+		if c >= 'A' && c <= 'Z' {
+
 			b[i] += 32
 		}
 	}
+
 	return string(b)
 }
 
@@ -34,12 +39,17 @@ func ASCIIToLower(s string) string {
 // are known to contain only ASCII and a faster conversion than is provided
 // by the standard `ToUpper()` is desired.
 func ASCIIToUpper(s string) string {
+
 	b := []byte(s)
-	for i := 0; i < len(b); i++ {
-		if b[i] >= 'a' && b[i] <= 'z' {
+
+	for i, c := range b {
+
+		if c >= 'a' && c <= 'z' {
+
 			b[i] -= 32
 		}
 	}
+
 	return string(b)
 }